arm disassembler support
This implements a basic disassembler for arm. It does not support thumb
at this time. This version punts on proerly formatting numbers based on
the libdis octal and hex options. It defaults to base 10 because that is
what the ARM PRM seems to suggest to sue. It also does not opt to format
certain mnemonics. I expect that will change as we get more experience
and exposure.