ld: Initial ARM cross-linker elfdump: Support ARM specifics Reviewed by: Robert Mustachi <rm@joyent.com>
This has a great chance of being very very broken, but it is sufficient to build
simple linux binaries, linked with this linker, and run them on raspbian.
Anything not directly useful to the kernel (or testing on raspbian) is missing.
There are probably unimplemented but important relocations.  TLS support is
entirely absent (and will assert if you try to use it).

Add knowledge of ARM specific section types, segment types, relocation types,
ELF flags, etc.

Support displaying the .ARM.attributes section.

rm agrees with Rich that the 24-bit relocation handling here is incredibly
suspicious, but for now we want to make forward progress on the port. So we've
opted to have future us curse past us because of our free-wheeling spirit about
this.