add a basic asm_linkage.h for arm
This adds support for the standard macros that are used commonly in most of our
asm files such as ENTRY() and SET_SIZE(). We currently do not implment the
macros necessary for the PROF and GPROF endpoints and actually punt on those
entirely. They may not even be necessary any more. We also do not yet have the
header files which this normaly includes, sys/stack.h and sys/trap.h. For the
mean time given that we don't even have anything like a trap or stack, these are
XXX'd out under an #if 0 section. We'll go back and add them in as they are
implemented.