Implement sys/machtypes.h Remove workarounds for non-functioning sys/types in sys/stack.h
This implements sys/machtypes.h which dictates the label_t type. An important
consideration is that we are currently opting to following the GNU variant of
the ABI with respect to r9. The ABI does not require all implements to treat r9
as a callee-saved register. Like the GNU/Linux conventions, we opt to do the
same thing.