Starting point for fakebop on arm Make a general boot_console.h interface Place files inside of platform-specific directories
This change goes through and changes a chunk of the plumbing around. More
specifically, it creates a platform specific directory for a given board and
starts to split up the interfaces between generic machine specific ones and
platform specific ones. Hence we've moved our initial locore and uart driver
into a board specific directory.

As a part of this we've begun the plumbing of fakebop. fakebop expects a boot
console to be potentially initialized. Because of that, we've bone through and
split that into a board specific interface as an almost primordial PSM. In the
future we should try and gather all of these into one coherent PSM that a given
board directory should implement.