Implement sys/machlock.h and sys/machsig.h for ARM
A note on sys/machlock.h: We basically have punted on the correctness of the
various ipl levels here. It's not immediately clear what exactly is going on in
the arm world. Because of that, there are sections which are XXX'd and under an
ifdef 0.

The correctness of machsig.h is probably suspicious. I suspect we will need
additional changes as time goes on to it, but most of the other platforms are
similar. We'll have to revisit it as we actually go and implement the various
signals and specific reason codes.