determine set of bootops we need and create stubs
This is a first pass at determining which of the boot operations we might need.
For now we have each one of them panic as we slowly proceed into kobj so we can
start to implement them based on what we know we need. Our stubs cause us to
panic so that's fine.