Add proto regset.h and ucontext.h
This adds a starting point for the regset.h and ucontext.h files. However it
leaves out and omits certain functionality.

Currently regset.h does not have any floating point information present in it.
When we add floating point information back inside of it we're going to have to
go back through ucontext.h and adjust the padding of the ucontext_t structure.