Get us to the point that kboj.o can build, but not link Add unoptimized versions of bcopy and strlen
Traditionally bcopy and strlen are implemented in asm on the target. Someday we
can go and waste a lot of time doing that, but not today.

This represents a chunk of work to get it so that we can start building parts of
krtld. Generally this involves adding a few #defines in the appropriate places.
All in all, this is progress even though we haven't implemented everything yet.
As a part of this, krtld needs a lot of common pieces so this will serve as the
reason to start getting our Makefiles in arm into better order.