Revamp Makefile system to match ON expectations Cleanup minor nits from a debug build
This is a bit of a whirlwind in our arm world. Here we're going through and
changing the build system and its mechanisms to more closely match the way that
i86pc/sun4[uv] work. This overlaps with our doing a debug build for the first
time, hence a lot of minor cleanup.

We had to rename misc/string.c to misc/arm_subr.c because string.o already
exists and the way the Makefile.files/Makefile.rules work they can only really
match a name in one directory. Oh well.