Add token ARM types for DDI headers
The DDI has a few types which are architecture dependent in common header files.
To deal with this, we've gone ahead and created basic types that look similar to
the x86 ones. As with many things, as we go to implement the DDI we will have a
better sense of what we need.