Set up to use u-boot for now.
For various reasons it appears that the rpi does not have a sane bootloader and
one that only reads in the first program header -- by all appearances it leaves
the .data section in the dust. We introduce a mapfile that puts data into a
rather high address for u-boot to be happy with it. To manually invoke it,
assuming a file called unix:

mmc rescan
fatload mmc 0:1 ${loadaddr} unix
bootelf unix