miniuart should inject \r on its own
Currently the kernel assumes that \n is sufficient to do both the equivalent of
a \n and a \r. It turns out that in the current miniuart drivers this is not the
case and therefore to match expecatations the miniuart will send a \r every time
it sees a \n.