kobj should not assume it can always mount the ramdisk
Currently kobj always assumes it can mount the ramdisk. It would take someone
royally messing up on i86pc to cause this to happen. However, it isn't as hard
for us to do that on ARM right now. In general we shouldn't just cast this to
void. If we ignore this failure we blow through and eventually hang because we
have a null bfs_ops.