add tools to build a boot_archive
This is a bit of a temporary hack. We need a boot_archive, but the means to
creating it are a bit up to us. For the moment we opt to create it via a shell
script we have embedded. Once we get a second module, then this will change
entirely.