Implement mod_modpath
Currently this takes an i86pc appraoch where by we are ignoring the specific
platform that this is. That said, this is probably not the long term goal as we
are likely to have lots of platform specific modules and implement this in a way
that is more common to sparc.