Implement kobj_reloc.c relocation support
This implements all of the support for relocationg kernel modules that krtld
needed that isn't a part of the generic doreloc.c shared pieces. We've opted not
to bother at all with tnf. tnf probes showing up will be unresolved symbols --
we don't really have much reason to implement it on ARM. SDT support is stubbed
out at this time.