Implement vmparam related header files
This goes through and clarifies our various *param definitions on ARMv6.
ARMv6 looks similar to amd64 and SPARC systems with respect to where
different componets of the address space are laid out. The stack grows
down, text and data are at the bottom of the address space. We have a
small red zone that exists between KERNELBASE and USERLIMIT. KERNELBASE
and in-kernel layout is similar to i386 because there is nothing like
the duplicate address spaces that SPARCv9 has in the ARM world.