Build and stub as much of genunix as you can Have SDT nop out, but not register probes
We need a large chunk of genunix and unix to make forward progress at
this time.  To facilitate that, we're trying to build as much of genunix
into it as we can and all of the missing and dependent functions are
instead coming from the fake stubs and the separate fake data. We will
have to continue this practice for unix as well. The goal is that we can
then piece by piece fix missing bits of files which don't build as well
as incrementally build our way up through the start up process.

As a side effect of this, we have a lot of stuff with SDT probes. As
such, we do the obvious thing where by we nop out the call site, but do
not record it at this time.