ASSERT_STACK_ALIGNED using wrong size local variable
ASSERT_STACK_ALIGNED works by putting a local variable of the
appropriate size onto the stack and then looking at its address. ARM
expects 8 byte stack alignment, but we were using a four byte value to
try and test that and thus breaking the way the test is designed. The
solution is to use an 8-byte value instead.