uts: primordial trap handling
Define the exception vector and map it at the "low" address (sadly, 0x0).

Handle all valid exceptions, and jump to entries in trap_table where handlers
are expected.

Define primordial handlers for early boot that merely make it clear that the
exception happened, and where practical sufficient details to identify it and
perhaps its cause.