Build and commanize prom emulation code Prepare gethrtime_unscaled for board specific impls. Fill in more of mlsetup()
Intel had its own basically fake version of the prom interface. We'd
like to leverage that for ARM. It's a bit janky to go reach into
uts/intel for files in Makefile.rules. Instead we've opted to copy it
into psm/promif/ so that way arm and intel can share the code. Note that
this change does not change anything for where uts/intel builds from.
That still exists pending someone testing the move.