fix constant assert in os/lgrp.c
uts/common/os/lgrp.c failed to build on ARM because of the following warning:

../../../common/os/lgrp.c: In function 'lgrp_mem_fini':
../../../common/os/lgrp.c:1444:3: error: comparison of unsigned expression >= 0
is always true [-Werror=type-limits]

The code in question was decrementing a reference effectively. Because the
assert can't catch overflow anyways it seemed like the better option is to move
the assert before we do the subtraction and change the compartor from '>= 0' to
simply '> 0'. Note that the operation is protected by holding cpu_lock so the
act of moving the assert with respect to the subtracton should be safe, given
the modification.