bcm2835: Use highvectors to work around u-boot bcm2835: Work around the fact that u-boot enables 'A' in the sctlr
u-boot does not properly load the exception table at the low vector. To work
around this we instead move it to the high vector in the mpafile and enable
highvecs. To add to the fun, u-boot also jumps to us with a SCTLR that is not
quite the documented post-reset default that we expect and see from QEMU. It
ends up making one major modification, it enables the 'A' bit in the SCTLR. This
enables strict alignment fault checking which ends up breaking us causing data
aborts in vnsprintf. Maybe some day we'll have a boot loader that works a bit
better for us...