no longer use special early-boot stacks
Run the primordial thread on t0stack from the outset, but reset sp when
we enter _locore_start (from which we'll never return).

Remove the special stacks for exceptional CPU modes in general
(primordial and otherwise).  The final revision of the trap code uses
srs/rfe and cps, and so does no stores prior to switching the CPU back
into supervisor mode and so onto the supervisor stack.  We leave the
exception mode's stack at #-1, such that any use of them will (in the
fullness of time), be immediately apparent.