primordial loader
This is a primordial loader. It's a bit janky, but it gets the job done. It adds
several new atag types and tools. The rough idea is that we have a blob which is
the boot_archive + unix glommed together with a header. The loader finds that,
sets up mappings for everything and goes to town. Ultimately the loader is going
to be a per-board deliverable. This work also motivated a revamped cache
interface.