Allow bootaux entries to override built in _etext, _edata
krtld relies on etending _etext and _edata up to some limit. On ARM,
the values of these which are in the binary are rather pathological.
Because of the ARM exception vector being at address 0xffff0000 the
link-editor changes what we think of as _edata and _etext and
kobj_setbrk ends up in a bad state. The simplest solution to this is to
allow someone to override these values in the bootaux vector like we do
here on ARM in combination with the loader.