make libgenunix.so an install target
The way that the genunix makefile was structured was a bit weird. While
libgenunix.so was implicitly an all target (and thus built genunix),
that wasn't exactly what we wanted. To make the rebuilt process a bit
faster and to better match what we want while in a bldenv environment,
though with no change to a nightly, we make libgenunix.so also a
dependency of the install starget.