stub vm_dep.h so we can start figuring out what it needs to have
We pretty much need the rest of the common unix objects to be ready to
go. To do that we need vm_dep.h. To know what it is we need inside
vm_dep.h the eaiest way is to stub it out and start adding functionality
so that we can actually build the other pieces of unix. This is that
stub.