Need a way to tell krtld about stext and sdata like etext and edata
Earlier we fixed things up such that we could tell krtld what the actual
ranges of the virtual text and data sections were. However, we only told
it where the end was. Unfortunately, the exception vector still threw
everything off and as such we somehow indicated in our binary that stext
was at the exception vector. Again, here we need to tell krtld an
alternate truth, just this time for the starting points and not just the
ending points.