overlay target table should actually hash the MAC address
Out of laziness we didn't come up with a hash. For now, we're using a
quick and dirty CRC32. It may not have ideal hashing properties, but
then again, the properties that we actually want are a bit odd in so far
as the first three bytes of the MAC will generally be fairly similar on
these networks.