Allow the svp plugin to try to connect Implement a libvarpd_panic() function
This allows the svp plugin to start to have more of a notion of
connecting to systems. In addition, it adds a panic function which
allows us to be a bit more explicit about where we're aborting and
include a more useful message. While generally all the state we care
about is in the core dump, this just adds a little bit of extra
information for a human to speed up the process.