Protype a new timer implementation
As we go further down the path of implementing reconnect logic and the
ilk, we need to have a sensible timer situation here. Because we don't
need a granularity larger than a second or particularly accurate timing,
we opt to use CLOCK_REALTIME instead of CLOCK_MONOTONIC, this will also
allow the svp plug-in to work in zones where the operator has not
allowed a high res timer.