Revamp the connectoin state machine Implement basic vl2/vl3 queries Use libvarpd_panic, stop using abort() Incorrectly determining svp timer expiration
This set of changes does a bunch of things. It revamps all of the logic
that we're using for our internal state machines for connections and
implements that new state machine. This better matches what we're doing
in other parts of SDC. In addition, clean up a few new bugs.