Properly set the invalid descriptor after close when backing off libidspace can't use VM_SLEEP allocations Add a standard bunyan logger for svp Validate most fields from the remote request
vmem_populate() requires that we pass VM_NOSLEEP in userland. This is
unfortunate, but seems to be a part of the current reality. This means
that at the moment we're changing around how we do the id_space_t
allocations and their ability to fail. Though that's a bit annoying, I'm
not sure what else we can really do at the moment, this is something
worth revisiting before we go back upstream.