Ignore spaces around the probe specifier name in dtrace's output.
When using one of the tracing options, such as -n, -i, -P, -f, if
the probe specifier contains any leading or trailing spaces, then
the first line of output produced by dtrace prints the probe name as
it is (along with the spaces). Moreover, if the probe also includes
a predicate or an action, then also there is a trailing space in the
probe name in the output. Following are couple of examples:

localhost$ sudo dtrace -n '  syscall:::entry '
dtrace: description '  syscall:::entry ' matched 482 probes

lcalhost$ sudo dtrace -n 'syscall:::entry {@num[probefunc] = count();}'
dtrace: description 'syscall:::entry ' matched 482 probes

This change skips the leading spaces in the probe name before
starting to process them. Also, when splitting the probe name
from the predicate and actions in compile_str(), it uses space
as one of the separators so that any trailing space after the
probe name will also be avoided in the output.

Also, add a test case (dtraceUtil/tst.TestProbeDescription.d.ksh) - which
checks that there are no leading or trailing spaces in the probe name
in dtrace's output when used using the -n option.